requirements.txt
adal==1.2.2
antlr4-python3-runtime==4.7.2
applicationinsights==0.11.9
argcomplete==1.10.0
asn1crypto==0.24.0
azure-batch==7.0.0
azure-cli==2.0.70
azure-cli-command-modules-nspkg==2.0.3
azure-cli-core==2.0.70
azure-cli-cosmosdb==0.2.11
azure-cli-nspkg==3.0.4
azure-cli-storage==2.4.3
azure-cli-telemetry==1.0.3
azure-common==1.1.23
azure-cosmos==3.1.0
azure-datalake-store==0.0.46
azure-functions-devops-build==0.0.22
azure-graphrbac==0.60.0
azure-keyvault==1.1.0
azure-mgmt-advisor==2.0.1
azure-mgmt-appconfiguration==0.1.0
azure-mgmt-applicationinsights==0.1.1
azure-mgmt-authorization==0.52.0
azure-mgmt-batch==6.0.0
azure-mgmt-batchai==2.0.0
azure-mgmt-billing==0.2.0
azure-mgmt-botservice==0.2.0
azure-mgmt-cdn==3.1.0
azure-mgmt-cognitiveservices==3.0.0
azure-mgmt-compute==6.0.0
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==1.5.0
azure-mgmt-containerregistry==3.0.0rc4
azure-mgmt-containerservice==5.3.0
azure-mgmt-cosmosdb==0.6.1
azure-mgmt-datalake-analytics==0.2.1
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==0.1.0
azure-mgmt-deploymentmanager==0.1.0
azure-mgmt-devtestlabs==2.2.0
azure-mgmt-dns==2.1.0
azure-mgmt-eventgrid==2.2.0
azure-mgmt-eventhub==2.6.0
azure-mgmt-hdinsight==1.1.0
azure-mgmt-imagebuilder==0.2.1
azure-mgmt-iotcentral==1.0.0
azure-mgmt-iothub==0.8.2
azure-mgmt-iothubprovisioningservices==0.2.0
azure-mgmt-keyvault==1.1.0
azure-mgmt-kusto==0.3.0
azure-mgmt-loganalytics==0.2.0
azure-mgmt-managedservices==1.0.0
azure-mgmt-managementgroups==0.2.0
azure-mgmt-maps==0.1.0
azure-mgmt-marketplaceordering==0.2.1
azure-mgmt-media==1.1.1
azure-mgmt-monitor==0.5.2
azure-mgmt-msi==0.2.0
azure-mgmt-netapp==0.5.0
azure-mgmt-network==3.0.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==0.3.1
azure-mgmt-privatedns==0.1.0
azure-mgmt-rdbms==1.9.0
azure-mgmt-recoveryservices==0.1.1
azure-mgmt-recoveryservicesbackup==0.1.2
azure-mgmt-redis==6.0.0
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.3.1
azure-mgmt-resource==2.2.0
azure-mgmt-search==2.1.0
azure-mgmt-security==0.1.0
azure-mgmt-servicebus==0.6.0
azure-mgmt-servicefabric==0.2.0
azure-mgmt-signalr==0.1.1
azure-mgmt-sql==0.12.0
azure-mgmt-sqlvirtualmachine==0.4.0
azure-mgmt-storage==4.0.0
azure-mgmt-trafficmanager==0.51.0
azure-mgmt-web==0.42.0
azure-multiapi-storage==0.2.4
azure-nspkg==3.0.2
azure-storage==0.36.0
azure-storage-blob==1.5.0
azure-storage-common==2.0.0
azure-storage-logging==0.5.1
azure-storage-queue==2.0.1
bcrypt==3.1.7
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.7
fabric==2.4.0
Flask==1.1.1
Flask-Menu==0.7.1
freeze==1.0.10
humanfriendly==4.18
idna==2.8
invoke==1.2.0
isodate==0.6.0
itsdangerous==1.1.0
javaproperties==0.5.1
Jinja2==2.10.1
jmespath==0.9.4
jsondiff==1.2.0
knack==0.6.3
MarkupSafe==1.1.1
mock==2.0.0
msrest==0.6.9
msrestazure==0.6.1
oauthlib==3.0.2
paramiko==2.6.0
pbr==5.4.1
portalocker==1.5.0
prompt-toolkit==2.0.9
psutil==5.6.3
pycparser==2.19
pydocumentdb==2.3.3
Pygments==2.4.2
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pypiwin32==223
pyreadline==2.1
python-dateutil==2.8.0
pytz==2019.1
pywin32==224
PyYAML==5.1.1
requests==2.22.0
requests-oauthlib==1.2.0
scp==0.13.2
six==1.12.0
sshtunnel==0.1.5
tabulate==0.8.3
urllib3==1.25.3
vsts==0.1.25
vsts-cd-manager==1.0.2
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.15.5
xmltodict==0.12.0
